fix: make OMP process lifecycle tests full-suite safe
Same class as grok-runtime: move the Symbol.for process.exit reaper onto
process-manager and reimport that module in the bound stress test so full-
suite shard 4 no longer times out on omp-runtime.
